requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.4
cryptography==3.2.1
Django==2.2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-modeltranslation==0.16
djangorestframework==3.12.1
fabric==2.5.0
fake==0.8
invoke==1.4.1
paramiko==2.7.2
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
wfastcgi==3.0.0
whitenoise==5.2.0
